move StudentRubricWidget to separate folder
